--- src/pages/AdminStationManagement.jsx
@@ -1,301 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { Container, Row, Col, Card, Button, Table, Modal, Form, Alert, Badge, Spinner } from 'react-bootstrap';
-import { stationAPI } from '../api/api';
-
-const AdminStationManagement = () => {
-  const [stations, setStations] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState('');
-  const [showModal, setShowModal] = useState(false);
-  const [editingStation, setEditingStation] = useState(null);
-  const [formData, setFormData] = useState({
-    name: '',
-    code: '',
-    latitude: '',
-    longitude: '',
-    address: '',
-    zone: '',
-    isActive: true
-  });
-
-  useEffect(() => {
-    fetchStations();
-  }, []);
-
-  const fetchStations = async () => {
-    try {
-      setLoading(true);
-      const response = await stationAPI.getAllStations();
-      setStations(Array.isArray(response.data) ? response.data : []);
-    } catch (error) {
-      setError('Failed to fetch stations');
-      console.error('Error fetching stations:', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      if (editingStation) {
-        await stationAPI.updateStation(editingStation._id, formData);
-      } else {
-        await stationAPI.createStation(formData);
-      }
-      setShowModal(false);
-      setEditingStation(null);
-      resetForm();
-      fetchStations();
-    } catch (error) {
-      setError(`Failed to ${editingStation ? 'update' : 'create'} station`);
-    }
-  };
-
-  const handleEdit = (station) => {
-    setEditingStation(station);
-    setFormData({
-      name: station.name || '',
-      code: station.code || '',
-      latitude: station.latitude || '',
-      longitude: station.longitude || '',
-      address: station.address || '',
-      zone: station.zone || '',
-      isActive: station.isActive !== false
-    });
-    setShowModal(true);
-  };
-
-  const handleDelete = async (stationId) => {
-    if (window.confirm('Are you sure you want to delete this station?')) {
-      try {
-        await stationAPI.deleteStation(stationId);
-        fetchStations();
-      } catch (error) {
-        setError('Failed to delete station');
-      }
-    }
-  };
-
-  const resetForm = () => {
-    setFormData({
-      name: '',
-      code: '',
-      latitude: '',
-      longitude: '',
-      address: '',
-      zone: '',
-      isActive: true
-    });
-  };
-
-  const handleCloseModal = () => {
-    setShowModal(false);
-    setEditingStation(null);
-    resetForm();
-  };
-
-  return (
-    <Container fluid className="p-4">
-      <Row className="mb-4">
-        <Col>
-          <div className="d-flex justify-content-between align-items-center">
-            <h2><i className="fas fa-subway me-2"></i>Station Management</h2>
-            <Button 
-              variant="primary" 
-              onClick={() => setShowModal(true)}
-            >
-              <i className="fas fa-plus me-2"></i>Add Station
-            </Button>
-          </div>
-        </Col>
-      </Row>
-
-      {error && (
-        <Row className="mb-3">
-          <Col>
-            <Alert variant="danger" dismissible onClose={() => setError('')}>
-              {error}
-            </Alert>
-          </Col>
-        </Row>
-      )}
-
-      <Row>
-        <Col>
-          <Card>
-            <Card.Header>
-              <h5 className="mb-0">All Stations ({stations.length})</h5>
-            </Card.Header>
-            <Card.Body>
-              {loading ? (
-                <div className="text-center py-4">
-                  <Spinner animation="border" />
-                  <p className="mt-2">Loading stations...</p>
-                </div>
-              ) : (
-                <Table responsive striped hover>
-                  <thead>
-                    <tr>
-                      <th>Name</th>
-                      <th>Code</th>
-                      <th>Zone</th>
-                      <th>Coordinates</th>
-                      <th>Status</th>
-                      <th>Actions</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {stations.map((station) => (
-                      <tr key={station._id}>
-                        <td>
-                          <strong>{station.name}</strong>
-                          {station.address && (
-                            <div className="text-muted small">{station.address}</div>
-                          )}
-                        </td>
-                        <td><code>{station.code}</code></td>
-                        <td>{station.zone || 'N/A'}</td>
-                        <td>
-                          {station.latitude && station.longitude ? (
-                            <small>
-                              {parseFloat(station.latitude).toFixed(4)}, {parseFloat(station.longitude).toFixed(4)}
-                            </small>
-                          ) : 'N/A'}
-                        </td>
-                        <td>
-                          <Badge bg={station.isActive !== false ? 'success' : 'danger'}>
-                            {station.isActive !== false ? 'Active' : 'Inactive'}
-                          </Badge>
-                        </td>
-                        <td>
-                          <Button
-                            variant="outline-primary"
-                            size="sm"
-                            className="me-2"
-                            onClick={() => handleEdit(station)}
-                          >
-                            <i className="fas fa-edit"></i>
-                          </Button>
-                          <Button
-                            variant="outline-danger"
-                            size="sm"
-                            onClick={() => handleDelete(station._id)}
-                          >
-                            <i className="fas fa-trash"></i>
-                          </Button>
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </Table>
-              )}
-            </Card.Body>
-          </Card>
-        </Col>
-      </Row>
-
-      {/* Add/Edit Modal */}
-      <Modal show={showModal} onHide={handleCloseModal} size="lg">
-        <Modal.Header closeButton>
-          <Modal.Title>
-            {editingStation ? 'Edit Station' : 'Add New Station'}
-          </Modal.Title>
-        </Modal.Header>
-        <Form onSubmit={handleSubmit}>
-          <Modal.Body>
-            <Row>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Station Name *</Form.Label>
-                  <Form.Control
-                    type="text"
-                    value={formData.name}
-                    onChange={(e) => setFormData({...formData, name: e.target.value})}
-                    required
-                  />
-                </Form.Group>
-              </Col>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Station Code *</Form.Label>
-                  <Form.Control
-                    type="text"
-                    value={formData.code}
-                    onChange={(e) => setFormData({...formData, code: e.target.value})}
-                    required
-                  />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Latitude</Form.Label>
-                  <Form.Control
-                    type="number"
-                    step="any"
-                    value={formData.latitude}
-                    onChange={(e) => setFormData({...formData, latitude: e.target.value})}
-                  />
-                </Form.Group>
-              </Col>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Longitude</Form.Label>
-                  <Form.Control
-                    type="number"
-                    step="any"
-                    value={formData.longitude}
-                    onChange={(e) => setFormData({...formData, longitude: e.target.value})}
-                  />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Form.Group className="mb-3">
-              <Form.Label>Address</Form.Label>
-              <Form.Control
-                as="textarea"
-                rows={2}
-                value={formData.address}
-                onChange={(e) => setFormData({...formData, address: e.target.value})}
-              />
-            </Form.Group>
-            <Row>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Zone</Form.Label>
-                  <Form.Control
-                    type="text"
-                    value={formData.zone}
-                    onChange={(e) => setFormData({...formData, zone: e.target.value})}
-                  />
-                </Form.Group>
-              </Col>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Check
-                    type="checkbox"
-                    label="Active Station"
-                    checked={formData.isActive}
-                    onChange={(e) => setFormData({...formData, isActive: e.target.checked})}
-                  />
-                </Form.Group>
-              </Col>
-            </Row>
-          </Modal.Body>
-          <Modal.Footer>
-            <Button variant="secondary" onClick={handleCloseModal}>
-              Cancel
-            </Button>
-            <Button variant="primary" type="submit">
-              {editingStation ? 'Update Station' : 'Create Station'}
-            </Button>
-          </Modal.Footer>
-        </Form>
-      </Modal>
-    </Container>
-  );
-};
-
-export default AdminStationManagement;
