--- src/pages/AdminUserManagement.jsx
@@ -1,375 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { Container, Row, Col, Card, Button, Table, Modal, Form, Alert, Badge, Spinner } from 'react-bootstrap';
-import { userAPI } from '../api/api';
-
-const AdminUserManagement = () => {
-  const [users, setUsers] = useState([]);
-  const [loading, setLoading] = useState(true);
-  const [error, setError] = useState('');
-  const [showModal, setShowModal] = useState(false);
-  const [editingUser, setEditingUser] = useState(null);
-  const [viewingUser, setViewingUser] = useState(null);
-  const [showViewModal, setShowViewModal] = useState(false);
-  const [formData, setFormData] = useState({
-    name: '',
-    email: '',
-    phone: '',
-    role: 'user',
-    isActive: true,
-    password: ''
-  });
-
-  useEffect(() => {
-    fetchUsers();
-  }, []);
-
-  const fetchUsers = async () => {
-    try {
-      setLoading(true);
-      const response = await userAPI.getAllUsers();
-      setUsers(Array.isArray(response.data) ? response.data : []);
-    } catch (error) {
-      setError('Failed to fetch users');
-      console.error('Error fetching users:', error);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    try {
-      if (editingUser) {
-        // Don't send password if it's empty during edit
-        const updateData = { ...formData };
-        if (!updateData.password) {
-          delete updateData.password;
-        }
-        await userAPI.updateUser(editingUser._id, updateData);
-      } else {
-        await userAPI.createUser(formData);
-      }
-      setShowModal(false);
-      setEditingUser(null);
-      resetForm();
-      fetchUsers();
-    } catch (error) {
-      setError(`Failed to ${editingUser ? 'update' : 'create'} user`);
-    }
-  };
-
-  const handleEdit = (user) => {
-    setEditingUser(user);
-    setFormData({
-      name: user.name || '',
-      email: user.email || '',
-      phone: user.phone || '',
-      role: user.role || 'user',
-      isActive: user.isActive !== false,
-      password: '' // Don't populate password for security
-    });
-    setShowModal(true);
-  };
-
-  const handleView = (user) => {
-    setViewingUser(user);
-    setShowViewModal(true);
-  };
-
-  const handleDelete = async (userId) => {
-    if (window.confirm('Are you sure you want to delete this user? This action cannot be undone.')) {
-      try {
-        await userAPI.deleteUser(userId);
-        fetchUsers();
-      } catch (error) {
-        setError('Failed to delete user');
-      }
-    }
-  };
-
-  const handleToggleStatus = async (userId, currentStatus) => {
-    try {
-      await userAPI.updateUser(userId, { isActive: !currentStatus });
-      fetchUsers();
-    } catch (error) {
-      setError('Failed to update user status');
-    }
-  };
-
-  const resetForm = () => {
-    setFormData({
-      name: '',
-      email: '',
-      phone: '',
-      role: 'user',
-      isActive: true,
-      password: ''
-    });
-  };
-
-  const handleCloseModal = () => {
-    setShowModal(false);
-    setEditingUser(null);
-    resetForm();
-  };
-
-  const getRoleBadge = (role) => {
-    const roleMap = {
-      'admin': 'danger',
-      'super_admin': 'dark',
-      'user': 'primary'
-    };
-    return roleMap[role] || 'secondary';
-  };
-
-  const formatDate = (dateString) => {
-    return new Date(dateString).toLocaleDateString();
-  };
-
-  return (
-    <Container fluid className="p-4">
-      <Row className="mb-4">
-        <Col>
-          <div className="d-flex justify-content-between align-items-center">
-            <h2><i className="fas fa-users me-2"></i>User Management</h2>
-            <Button 
-              variant="primary" 
-              onClick={() => setShowModal(true)}
-            >
-              <i className="fas fa-plus me-2"></i>Add User
-            </Button>
-          </div>
-        </Col>
-      </Row>
-
-      {error && (
-        <Row className="mb-3">
-          <Col>
-            <Alert variant="danger" dismissible onClose={() => setError('')}>
-              {error}
-            </Alert>
-          </Col>
-        </Row>
-      )}
-
-      <Row>
-        <Col>
-          <Card>
-            <Card.Header>
-              <h5 className="mb-0">All Users ({users.length})</h5>
-            </Card.Header>
-            <Card.Body>
-              {loading ? (
-                <div className="text-center py-4">
-                  <Spinner animation="border" />
-                  <p className="mt-2">Loading users...</p>
-                </div>
-              ) : (
-                <Table responsive striped hover>
-                  <thead>
-                    <tr>
-                      <th>Name</th>
-                      <th>Email</th>
-                      <th>Phone</th>
-                      <th>Role</th>
-                      <th>Status</th>
-                      <th>Joined</th>
-                      <th>Actions</th>
-                    </tr>
-                  </thead>
-                  <tbody>
-                    {users.map((user) => (
-                      <tr key={user._id}>
-                        <td><strong>{user.name || 'N/A'}</strong></td>
-                        <td>{user.email || 'N/A'}</td>
-                        <td>{user.phone || 'N/A'}</td>
-                        <td>
-                          <Badge bg={getRoleBadge(user.role)}>
-                            {user.role || 'user'}
-                          </Badge>
-                        </td>
-                        <td>
-                          <Badge bg={user.isActive ? 'success' : 'danger'}>
-                            {user.isActive ? 'Active' : 'Inactive'}
-                          </Badge>
-                        </td>
-                        <td>{user.createdAt ? formatDate(user.createdAt) : 'N/A'}</td>
-                        <td>
-                          <Button
-                            variant="outline-info"
-                            size="sm"
-                            className="me-2"
-                            onClick={() => handleView(user)}
-                          >
-                            <i className="fas fa-eye"></i>
-                          </Button>
-                          <Button
-                            variant="outline-primary"
-                            size="sm"
-                            className="me-2"
-                            onClick={() => handleEdit(user)}
-                          >
-                            <i className="fas fa-edit"></i>
-                          </Button>
-                          <Button
-                            variant={user.isActive ? "outline-warning" : "outline-success"}
-                            size="sm"
-                            className="me-2"
-                            onClick={() => handleToggleStatus(user._id, user.isActive)}
-                          >
-                            <i className={`fas fa-${user.isActive ? 'ban' : 'check'}`}></i>
-                          </Button>
-                          <Button
-                            variant="outline-danger"
-                            size="sm"
-                            onClick={() => handleDelete(user._id)}
-                          >
-                            <i className="fas fa-trash"></i>
-                          </Button>
-                        </td>
-                      </tr>
-                    ))}
-                  </tbody>
-                </Table>
-              )}
-            </Card.Body>
-          </Card>
-        </Col>
-      </Row>
-
-      {/* Add/Edit Modal */}
-      <Modal show={showModal} onHide={handleCloseModal} size="lg">
-        <Modal.Header closeButton>
-          <Modal.Title>
-            {editingUser ? 'Edit User' : 'Add New User'}
-          </Modal.Title>
-        </Modal.Header>
-        <Form onSubmit={handleSubmit}>
-          <Modal.Body>
-            <Row>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Full Name *</Form.Label>
-                  <Form.Control
-                    type="text"
-                    value={formData.name}
-                    onChange={(e) => setFormData({...formData, name: e.target.value})}
-                    required
-                    placeholder="Enter full name"
-                  />
-                </Form.Group>
-              </Col>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Email *</Form.Label>
-                  <Form.Control
-                    type="email"
-                    value={formData.email}
-                    onChange={(e) => setFormData({...formData, email: e.target.value})}
-                    required
-                    placeholder="Enter email address"
-                  />
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Phone</Form.Label>
-                  <Form.Control
-                    type="tel"
-                    value={formData.phone}
-                    onChange={(e) => setFormData({...formData, phone: e.target.value})}
-                    placeholder="Enter phone number"
-                  />
-                </Form.Group>
-              </Col>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>Role</Form.Label>
-                  <Form.Select
-                    value={formData.role}
-                    onChange={(e) => setFormData({...formData, role: e.target.value})}
-                  >
-                    <option value="user">User</option>
-                    <option value="admin">Admin</option>
-                    <option value="super_admin">Super Admin</option>
-                  </Form.Select>
-                </Form.Group>
-              </Col>
-            </Row>
-            <Row>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Label>
-                    Password {editingUser ? '(Leave blank to keep current)' : '*'}
-                  </Form.Label>
-                  <Form.Control
-                    type="password"
-                    value={formData.password}
-                    onChange={(e) => setFormData({...formData, password: e.target.value})}
-                    required={!editingUser}
-                    placeholder={editingUser ? "Enter new password" : "Enter password"}
-                  />
-                </Form.Group>
-              </Col>
-              <Col md={6}>
-                <Form.Group className="mb-3">
-                  <Form.Check
-                    type="checkbox"
-                    label="Active User"
-                    checked={formData.isActive}
-                    onChange={(e) => setFormData({...formData, isActive: e.target.checked})}
-                  />
-                </Form.Group>
-              </Col>
-            </Row>
-          </Modal.Body>
-          <Modal.Footer>
-            <Button variant="secondary" onClick={handleCloseModal}>
-              Cancel
-            </Button>
-            <Button variant="primary" type="submit">
-              {editingUser ? 'Update User' : 'Create User'}
-            </Button>
-          </Modal.Footer>
-        </Form>
-      </Modal>
-
-      {/* View Modal */}
-      <Modal show={showViewModal} onHide={() => setShowViewModal(false)} size="lg">
-        <Modal.Header closeButton>
-          <Modal.Title>User Details</Modal.Title>
-        </Modal.Header>
-        <Modal.Body>
-          {viewingUser && (
-            <Row>
-              <Col md={6}>
-                <h6>Basic Information</h6>
-                <p><strong>User ID:</strong> {viewingUser._id}</p>
-                <p><strong>Name:</strong> {viewingUser.name || 'N/A'}</p>
-                <p><strong>Email:</strong> {viewingUser.email || 'N/A'}</p>
-                <p><strong>Phone:</strong> {viewingUser.phone || 'N/A'}</p>
-                <p><strong>Role:</strong> <Badge bg={getRoleBadge(viewingUser.role)}>{viewingUser.role || 'user'}</Badge></p>
-              </Col>
-              <Col md={6}>
-                <h6>Account Details</h6>
-                <p><strong>Status:</strong> <Badge bg={viewingUser.isActive ? 'success' : 'danger'}>{viewingUser.isActive ? 'Active' : 'Inactive'}</Badge></p>
-                <p><strong>Joined:</strong> {viewingUser.createdAt ? formatDate(viewingUser.createdAt) : 'N/A'}</p>
-                <p><strong>Last Updated:</strong> {viewingUser.updatedAt ? formatDate(viewingUser.updatedAt) : 'N/A'}</p>
-                <p><strong>Wallet Balance:</strong> ₹{viewingUser.walletBalance || 0}</p>
-              </Col>
-            </Row>
-          )}
-        </Modal.Body>
-        <Modal.Footer>
-          <Button variant="secondary" onClick={() => setShowViewModal(false)}>
-            Close
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    </Container>
-  );
-};
-
-export default AdminUserManagement;
