+++ src/pages/admin/AdminLogin.jsx
@@ -1,15 +1,15 @@
 import React, { useState } from 'react';
-import { Container, Row, Col, Card, Form, Button, Alert, Spinner } from 'react-bootstrap';
-import { useNavigate, Link } from 'react-router-dom';
+import { Container, Card, Form, Button, Alert, Spinner, Row, Col } from 'react-bootstrap';
+import { useNavigate } from 'react-router-dom';
 import { adminAPI } from '../../api/api';
 
 const AdminLogin = () => {
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState('');
   const [formData, setFormData] = useState({
     email: '',
     password: ''
   });
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState('');
   const navigate = useNavigate();
 
   const handleChange = (e) => {
@@ -23,41 +23,56 @@ const AdminLogin = () => {
     e.preventDefault();
     setLoading(true);
     setError('');
-
+    
     try {
-      const response = await adminAPI.login(formData);
-      
-      if (response.data && response.data.token) {
-        // Store admin token
-        localStorage.setItem('adminToken', response.data.token);
-        localStorage.setItem('adminUser', JSON.stringify(response.data.admin));
+      const response = await adminAPI.login({
+        email: formData.email,
+        password: formData.password
+      });
+
+      if (response.data.success) {
+        const { token, user } = response.data.data;
+        
+        console.log('Login successful, received data:', { token: token ? 'Token received' : 'No token', user });
+        
+        // Store admin token and user info
+        localStorage.setItem('adminToken', token);
+        localStorage.setItem('adminUser', JSON.stringify(user));
+        
+        console.log('Token stored in localStorage:', localStorage.getItem('adminToken') ? 'Yes' : 'No');
+        console.log('User stored in localStorage:', localStorage.getItem('adminUser'));
         
         // Navigate to admin dashboard
         navigate('/admin/dashboard');
       } else {
-        setError('Invalid response from server');
+        setError(response.data.message || 'Login failed');
       }
     } catch (error) {
-      console.error('Admin login error:', error);
-      setError(error.response?.data?.message || 'Login failed. Please try again.');
+      console.error('Login error:', error);
+      console.error('Error response:', error.response?.data);
+      setError(error.response?.data?.message || 'Login failed. Please check your credentials.');
     } finally {
       setLoading(false);
     }
   };
 
   return (
-    <div className="min-vh-100 d-flex align-items-center bg-light">
+    <div style={{
+      minHeight: '100vh',
+      background: 'linear-gradient(135deg, #667eea 0%, #764ba2 100%)',
+      display: 'flex',
+      alignItems: 'center',
+      justifyContent: 'center',
+      padding: '20px'
+    }}>
       <Container>
         <Row className="justify-content-center">
           <Col md={6} lg={4}>
-            <Card className="shadow">
+            <Card className="shadow-lg border-0" style={{ borderRadius: '12px' }}>
               <Card.Body className="p-4">
                 <div className="text-center mb-4">
-                  <h2 className="text-primary">
-                    <i className="fas fa-shield-alt me-2"></i>
-                    Admin Login
-                  </h2>
-                  <p className="text-muted">Tapido Administration Panel</p>
+                  <h2 className="text-primary mb-2">Admin Portal</h2>
+                  <p className="text-muted">SmartMetroCard Administration</p>
                 </div>
 
                 {error && (
@@ -68,14 +83,15 @@ const AdminLogin = () => {
 
                 <Form onSubmit={handleSubmit}>
                   <Form.Group className="mb-3">
-                    <Form.Label>Email Address</Form.Label>
+                    <Form.Label>Email</Form.Label>
                     <Form.Control
                       type="email"
                       name="email"
                       value={formData.email}
                       onChange={handleChange}
+                      placeholder="admin@smartmetrocard.com"
                       required
-                      placeholder="Enter admin email"
+                      size="lg"
                     />
                   </Form.Group>
 
@@ -86,16 +102,20 @@ const AdminLogin = () => {
                       name="password"
                       value={formData.password}
                       onChange={handleChange}
+                      placeholder="Enter your password"
                       required
-                      placeholder="Enter password"
+                      minLength={6}
+                      size="lg"
                     />
                   </Form.Group>
 
                   <Button
                     type="submit"
                     variant="primary"
-                    className="w-100 mb-3"
+                    size="lg"
+                    className="w-100"
                     disabled={loading}
+                    style={{ borderRadius: '8px', fontWeight: '500' }}
                   >
                     {loading ? (
                       <>
@@ -108,10 +128,20 @@ const AdminLogin = () => {
                   </Button>
                 </Form>
 
+                <div className="text-center mt-4">
+                  <small className="text-muted">
+                    Authorized personnel only. All activities are logged.
+                  </small>
+                </div>
+
                 <div className="text-center">
-                  <Link to="/admin/register" className="text-decoration-none">
-                    Need an admin account? Register here
-                  </Link>
+                  <Button
+                    variant="link"
+                    onClick={() => navigate('/admin/register')}
+                    className="text-decoration-none"
+                  >
+                    Don't have an account? Register here
+                  </Button>
                 </div>
               </Card.Body>
             </Card>
