+++ src/pages/admin/AdminRegister.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
-import { Container, Row, Col, Card, Form, Button, Alert, Spinner } from 'react-bootstrap';
-import { useNavigate, Link } from 'react-router-dom';
+import { Container, Row, Col, Card, Form, Button, Alert } from 'react-bootstrap';
+import { useNavigate } from 'react-router-dom';
 import { adminAPI } from '../../api/api';
 
 const AdminRegister = () => {
@@ -9,10 +9,12 @@ const AdminRegister = () => {
     email: '',
     password: '',
     confirmPassword: '',
-    role: 'admin'
+    role: 'admin',
+    phoneNumber: ''
   });
   const [loading, setLoading] = useState(false);
   const [error, setError] = useState('');
+  const [success, setSuccess] = useState('');
   const navigate = useNavigate();
 
   const handleChange = (e) => {
@@ -24,36 +26,51 @@ const AdminRegister = () => {
 
   const handleSubmit = async (e) => {
     e.preventDefault();
-    setLoading(true);
     setError('');
+    setSuccess('');
 
-    // Validate passwords match
-    if (formData.password !== formData.confirmPassword) {
-      setError('Passwords do not match');
-      setLoading(false);
+    // Validation
+    if (!formData.name || !formData.email || !formData.password) {
+      setError('Name, email, and password are required');
       return;
     }
+
+    if (formData.password !== formData.confirmPassword) {
+      setError('Passwords do not match');
+      return;
+    }
+
+    if (formData.password.length < 6) {
+      setError('Password must be at least 6 characters long');
+      return;
+    }
+
+    setLoading(true);
 
     try {
       const response = await adminAPI.register({
         name: formData.name,
         email: formData.email,
         password: formData.password,
-        role: formData.role
+        role: formData.role,
+        phoneNumber: formData.phoneNumber
       });
-      
-      if (response.data && response.data.token) {
-        // Store admin token
+
+      if (response.success) {
+        // Store admin token and user info
         localStorage.setItem('adminToken', response.data.token);
-        localStorage.setItem('adminUser', JSON.stringify(response.data.admin));
+        localStorage.setItem('adminUser', JSON.stringify(response.data.user));
         
-        // Navigate to admin dashboard
-        navigate('/admin/dashboard');
+        setSuccess('Admin account created successfully! Redirecting to dashboard...');
+        setTimeout(() => {
+          navigate('/admin/dashboard');
+        }, 2000);
       } else {
-        setError('Invalid response from server');
+        setError(response.message || 'Registration failed');
       }
     } catch (error) {
-      console.error('Admin registration error:', error);
+      console.error('Registration error:', error);
+      console.error('Error response:', error.response?.data);
       setError(error.response?.data?.message || 'Registration failed. Please try again.');
     } finally {
       setLoading(false);
@@ -61,116 +78,125 @@ const AdminRegister = () => {
   };
 
   return (
-    <div className="min-vh-100 d-flex align-items-center bg-light">
-      <Container>
-        <Row className="justify-content-center">
-          <Col md={6} lg={5}>
-            <Card className="shadow">
-              <Card.Body className="p-4">
-                <div className="text-center mb-4">
-                  <h2 className="text-primary">
-                    <i className="fas fa-user-shield me-2"></i>
-                    Admin Registration
-                  </h2>
-                  <p className="text-muted">Create Tapido Admin Account</p>
-                </div>
+    <Container fluid className="min-vh-100 d-flex align-items-center justify-content-center bg-light">
+      <Row className="w-100 justify-content-center">
+        <Col md={6} lg={4}>
+          <Card className="shadow">
+            <Card.Body className="p-4">
+              <div className="text-center mb-4">
+                <h3 className="text-primary mb-2">Admin Registration</h3>
+                <p className="text-muted">Create a new admin account</p>
+              </div>
 
-                {error && (
-                  <Alert variant="danger" className="mb-3">
-                    {error}
-                  </Alert>
-                )}
+              {error && <Alert variant="danger">{error}</Alert>}
+              {success && <Alert variant="success">{success}</Alert>}
 
-                <Form onSubmit={handleSubmit}>
-                  <Form.Group className="mb-3">
-                    <Form.Label>Full Name</Form.Label>
-                    <Form.Control
-                      type="text"
-                      name="name"
-                      value={formData.name}
-                      onChange={handleChange}
-                      required
-                      placeholder="Enter full name"
-                    />
-                  </Form.Group>
+              <Form onSubmit={handleSubmit}>
+                <Form.Group className="mb-3">
+                  <Form.Label>Full Name</Form.Label>
+                  <Form.Control
+                    type="text"
+                    name="name"
+                    value={formData.name}
+                    onChange={handleChange}
+                    placeholder="Enter your full name"
+                    required
+                  />
+                </Form.Group>
 
-                  <Form.Group className="mb-3">
-                    <Form.Label>Email Address</Form.Label>
-                    <Form.Control
-                      type="email"
-                      name="email"
-                      value={formData.email}
-                      onChange={handleChange}
-                      required
-                      placeholder="Enter email address"
-                    />
-                  </Form.Group>
+                <Form.Group className="mb-3">
+                  <Form.Label>Email Address</Form.Label>
+                  <Form.Control
+                    type="email"
+                    name="email"
+                    value={formData.email}
+                    onChange={handleChange}
+                    placeholder="Enter your email"
+                    required
+                  />
+                </Form.Group>
 
-                  <Form.Group className="mb-3">
-                    <Form.Label>Role</Form.Label>
-                    <Form.Select
-                      name="role"
-                      value={formData.role}
-                      onChange={handleChange}
-                      required
-                    >
-                      <option value="admin">Admin</option>
-                      <option value="super_admin">Super Admin</option>
-                    </Form.Select>
-                  </Form.Group>
+                <Form.Group className="mb-3">
+                  <Form.Label>Phone Number (Optional)</Form.Label>
+                  <Form.Control
+                    type="tel"
+                    name="phoneNumber"
+                    value={formData.phoneNumber}
+                    onChange={handleChange}
+                    placeholder="Enter your phone number"
+                  />
+                </Form.Group>
 
-                  <Form.Group className="mb-3">
-                    <Form.Label>Password</Form.Label>
-                    <Form.Control
-                      type="password"
-                      name="password"
-                      value={formData.password}
-                      onChange={handleChange}
-                      required
-                      placeholder="Enter password"
-                    />
-                  </Form.Group>
-
-                  <Form.Group className="mb-4">
-                    <Form.Label>Confirm Password</Form.Label>
-                    <Form.Control
-                      type="password"
-                      name="confirmPassword"
-                      value={formData.confirmPassword}
-                      onChange={handleChange}
-                      required
-                      placeholder="Confirm password"
-                    />
-                  </Form.Group>
-
-                  <Button
-                    type="submit"
-                    variant="primary"
-                    className="w-100 mb-3"
-                    disabled={loading}
+                <Form.Group className="mb-3">
+                  <Form.Label>Role</Form.Label>
+                  <Form.Select
+                    name="role"
+                    value={formData.role}
+                    onChange={handleChange}
                   >
-                    {loading ? (
-                      <>
-                        <Spinner animation="border" size="sm" className="me-2" />
-                        Creating Account...
-                      </>
-                    ) : (
-                      'Create Admin Account'
-                    )}
-                  </Button>
-                </Form>
+                    <option value="admin">Admin</option>
+                    <option value="super_admin">Super Admin</option>
+                  </Form.Select>
+                </Form.Group>
+
+                <Form.Group className="mb-3">
+                  <Form.Label>Password</Form.Label>
+                  <Form.Control
+                    type="password"
+                    name="password"
+                    value={formData.password}
+                    onChange={handleChange}
+                    placeholder="Enter your password"
+                    required
+                  />
+                  <Form.Text className="text-muted">
+                    Password must be at least 6 characters long
+                  </Form.Text>
+                </Form.Group>
+
+                <Form.Group className="mb-4">
+                  <Form.Label>Confirm Password</Form.Label>
+                  <Form.Control
+                    type="password"
+                    name="confirmPassword"
+                    value={formData.confirmPassword}
+                    onChange={handleChange}
+                    placeholder="Confirm your password"
+                    required
+                  />
+                </Form.Group>
+
+                <Button
+                  variant="primary"
+                  type="submit"
+                  className="w-100 mb-3"
+                  disabled={loading}
+                >
+                  {loading ? (
+                    <>
+                      <span className="spinner-border spinner-border-sm me-2" role="status" aria-hidden="true"></span>
+                      Creating Account...
+                    </>
+                  ) : (
+                    'Create Admin Account'
+                  )}
+                </Button>
 
                 <div className="text-center">
-                  <Link to="/admin/login" className="text-decoration-none">
-                    Already have an account? Sign in
-                  </Link>
+                  <Button
+                    variant="link"
+                    onClick={() => navigate('/admin/login')}
+                    className="text-decoration-none"
+                  >
+                    Already have an account? Login here
+                  </Button>
                 </div>
-              </Card.Body>
-            </Card>
-          </Col>
-        </Row>
-      </Container>
-    </div>
+              </Form>
+            </Card.Body>
+          </Card>
+        </Col>
+      </Row>
+    </Container>
   );
 };
